Support rubocop versions > 38
The TargetRubyVersion config in ruby has gone some heavy overhaul since
it was first introduced in rubocop 0.36.0, and now there is a
target_ruby_version method available directly from the RuboCop::Config
class as of 0.41.0.

This makes implementation added here redundant for newer versions of
RuboCop, and also causes the tests to fail when resolving the
TargetRubyVersion through it.  It shall be used for rubocop versions
less than 0.41.0 until those are not supported, but on anything greater
than that, we shall call out to the RuboCop::Config API directly for
resolution of that (it will use RuboCop::Config::DEFAULT_RUBY_VERSION if
it is not able to determine one, avoiding errors).