Support rubocop 0.47.0+ and previous versions
In rubocop 0.47.0, the `RuboCop::Cop::Registry` was added as the new way
of handling passing around a collection of Cops to other portions of the
code base.  See https://github.com/bbatsov/rubocop/pull/3837

This change, as a result, made it so that initializing
`RuboCop::Cop::Team` required you passing in a `Registry`, and not an
Array of cops (which was previously supported`.

Because this change is not backwards compatible, we need to support two
ways of intializing the `RuboCop::Cop::Team`:

1. By passing it an array of cops:  `RuboCop::Cop::Cop.all`
2. By passing it a registry of cops:  `RuboCop::Cop::Registry.new cops`

And this is dependent on the version, so the `all_cops` methods will be
conditionally created based on the VERSION of rubocop that supports
either call.